fix: remove unnecessary as u32 cast in expect_actual test

+++ src/cli/expect_actual_tests.rs
@@ -24,7 +24,7 @@ fn first_id_no_panic() {
     let src = "class Foo";
     let tree = parse_kotlin(src).unwrap();
     let root = tree.root_node();
-    if let Some(decl) = root.child(0 as u32) {
+    if let Some(decl) = root.child(0) {
         let _id = first_id(&decl, src);
     }
 }
